+++ voucher-manager.php
@@ -3,7 +3,7 @@
  * Plugin Name:       Voucher Manager
  * Plugin URI:        https://github.com/mjs512/voucher-manager
  * Description:       Professional One-Time Code Management for WordPress.
- * Version:           1.0.0
+ * Version:           1.0.1
  * Requires at least: 6.5
  * Requires PHP:      8.1
  * Author:            MJS-Productions
@@ -22,7 +22,7 @@ if ( ! defined( 'ABSPATH' ) ) {
 	exit;
 }
 
-define( 'VOUCHER_MANAGER_VERSION', '1.0.0' );
+define( 'VOUCHER_MANAGER_VERSION', '1.0.1' );
 define( 'VOUCHER_MANAGER_DATABASE_VERSION', '2' );
 define( 'VOUCHER_MANAGER_FILE', __FILE__ );
 define( 'VOUCHER_MANAGER_PATH', plugin_dir_path( __FILE__ ) );